Refactored AbstractCPP Should be easier to manage
