+++ AbstractCPP/include/AbstractCPP.hpp
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "abstractcpp/date.hpp"
+#include "abstractcpp/string.hpp"
+#include "abstractcpp/exceptions.hpp"
