+++ AbstractCPP/include/abstractcpp/date.hpp
@@ -1,26 +1,9 @@
 #pragma once
 
 #include <string>
-#include <vector>
 #include <chrono>
 
 namespace abstractcpp {
-	class string
-	{
-	private:
-		std::string internalString;
-
-	public:
-		string(std::string initialString);
-
-		std::vector<std::string> split(const char* splitter);
-		bool contains(std::string searchString);
-		bool startsWith(std::string comparator);
-		bool endsWith(std::string comparator);
-
-		std::string str() { return internalString; }
-	};
-
 	class Date
 	{
 	private:
