+++ AbstractCPP/include/abstractcpp/exceptions.hpp
